pyproject.toml
[project]
name = "aitk"
version = "0.1.0"
description = "AI Toolkit CLI"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "firecrawl-py>=1.0.0",
    "playwright>=1.40.0",
    "imageio[ffmpeg]>=2.34.0",
]

[project.scripts]
aitk = "aitk:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aitk"]

[tool.ruff]
line-length = 100
target-version = "py312"
